+++ src/pages/Popup/__tests-helpers__/functions.tsx
@@ -2,6 +2,8 @@ import { render, screen } from '@testing-library/react';
 import { ReactWrapper } from 'enzyme';
 import { ReactElement } from 'react';
 
+// todo this entire folder ignored from production build?
+
 export const getButtonByText = (wrapper: ReactWrapper, btnText: string) =>
   wrapper.findWhere((node) => {
     return node.type() === 'button' && node.text() === btnText;
@@ -14,4 +16,16 @@ export const renderComponentAndExpect = (
   render(component);
   const element = screen.getByText(expectString);
   expect(element).toBeInTheDocument();
+};
+
+// will be useful if we ever expand on what attributes we can match
+export const generateFakeTab = (newFakeTabArgs: Partial<chrome.tabs.Tab>) => {
+  return Object.assign(
+    {
+      id: Math.floor(Math.random() * 100000 + 1),
+      title: 'my fake tab',
+      url: 'http://google.ca',
+    },
+    newFakeTabArgs
+  );
 };